Adjust to handle higlhighted rows
Some chronorace result pages have rows that highlight on hover
(specifically when there is a link to a more detailed info page per
athlete). This breaks our matching on the rows as the tag is more
elaborate than just the one class. Leave an open ending when matching
to handle these cases too. Also add a test case.