Fix #1: Find next pages on a Dutch page too
The code already has special considerations for parsing the athletes out
of a Dutch page, but failed to find the next pages to parse all the
results. It was only looking for "Page: [1][2]..." while in Dutch it
says "Bladzijde" there. Adjust it to also consider that situation. Hope
the few tests catch any regressions.